+++ pkg/nodeconfig/nodeconfig.go
@@ -21,7 +21,6 @@ import (
 	ctrl "sigs.k8s.io/controller-runtime"
 	crclientcfg "sigs.k8s.io/controller-runtime/pkg/client/config"
 
-	"github.com/openshift/windows-machine-config-operator/pkg/annotations"
 	"github.com/openshift/windows-machine-config-operator/pkg/cluster"
 	"github.com/openshift/windows-machine-config-operator/pkg/instances"
 	"github.com/openshift/windows-machine-config-operator/pkg/metadata"
@@ -219,7 +218,7 @@ func (nc *nodeConfig) Configure() error {
 
 // applyAnnotations applies all the given annotations and updates the Node object in NodeConfig
 func (nc *nodeConfig) applyAnnotations(annotationsToApply map[string]string) error {
-	patchData, err := annotations.GenerateAddPatch(annotationsToApply)
+	patchData, err := metadata.GenerateAddPatch(annotationsToApply)
 	if err != nil {
 		return err
 	}
@@ -373,7 +372,7 @@ func (nc *nodeConfig) Deconfigure() error {
 	}
 
 	// Clear the version annotation from the node object to indicate the node is not configured
-	patchData, err := annotations.GenerateRemovePatch([]string{metadata.VersionAnnotation})
+	patchData, err := metadata.GenerateRemovePatch([]string{metadata.VersionAnnotation})
 	if err != nil {
 		return errors.Wrapf(err, "error creating version annotation remove request")
 	}
